% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{BALANCE_SHEET_CONFIG}
\alias{BALANCE_SHEET_CONFIG}
\title{Balance Sheet Data Configuration}
\format{
A list containing configuration parameters for balance sheet data fetching.
}
\usage{
BALANCE_SHEET_CONFIG
}
\description{
Configuration object for fetching quarterly balance sheet data from Alpha Vantage.
}
\keyword{datasets}
