% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_baseline.R
\name{calculate_baseline}
\alias{calculate_baseline}
\title{Calculate Baseline Indices for Anomaly Detection}
\usage{
calculate_baseline(i, n, lookback, lookahead)
}
\arguments{
\item{i}{Integer position in the time series for which to calculate baseline indices}

\item{n}{Integer length of the time series}

\item{lookback}{Integer number of periods to look back for baseline calculation}

\item{lookahead}{Integer number of periods to look ahead for baseline calculation}
}
\value{
Integer vector of indices to use for baseline calculation
}
\description{
Calculates which indices should be used for baseline calculation for a given position,
excluding the current point and immediate neighbors to prevent contamination.
}
