% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{CASH_FLOW_CONFIG}
\alias{CASH_FLOW_CONFIG}
\title{Cash Flow Data Configuration}
\format{
A list containing configuration parameters for cash flow data fetching.
}
\usage{
CASH_FLOW_CONFIG
}
\description{
Configuration object for fetching quarterly cash flow data from Alpha Vantage.
}
\keyword{datasets}
