% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ticker_results.R
\name{combine_ticker_results}
\alias{combine_ticker_results}
\title{Combine ticker results into a single dataframe}
\usage{
combine_ticker_results(results_list, tickers)
}
\arguments{
\item{results_list}{List of tibbles with ticker data (may contain NULL entries)}

\item{tickers}{Character vector. Original vector of equity symbols}
}
\value{
A tibble with combined daily adjusted price data for all successful tickers
}
\description{
Combine ticker results into a single dataframe
}
\keyword{internal}
