% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_logger.R
\name{create_pipeline_log}
\alias{create_pipeline_log}
\title{Initialize Pipeline Log}
\usage{
create_pipeline_log()
}
\value{
tibble: Empty log dataframe
}
\description{
Creates an empty pipeline log dataframe.
}
\keyword{internal}
