% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_weekly_dates.R
\name{derive_weekly_dates}
\alias{derive_weekly_dates}
\title{Derive Weekly Observation Dates from Daily Price Data}
\usage{
derive_weekly_dates(price_data, n_weeks = 52L)
}
\arguments{
\item{price_data}{Tibble with at least a \code{date} column of daily prices}

\item{n_weeks}{Integer. Number of most recent weeks to return}
}
\value{
Date vector of the last trading day per week, most recent first
}
\description{
Derive Weekly Observation Dates from Daily Price Data
}
