% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_data_changes.R
\name{detect_data_changes}
\alias{detect_data_changes}
\title{Detect if New Data Contains Changes}
\usage{
detect_data_changes(existing_data, new_data, date_column)
}
\arguments{
\item{existing_data}{data.frame: Previously stored data (can be NULL)}

\item{new_data}{data.frame: Newly fetched data}

\item{date_column}{character: Column name to use for comparison (e.g., "fiscalDateEnding")}
}
\value{
list: has_changes (logical), new_records_count (integer), latest_date (Date or NULL)
}
\description{
Compares new data to existing data to determine if there are actual changes.
}
\keyword{internal}
