% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{EARNINGS_CONFIG}
\alias{EARNINGS_CONFIG}
\title{Earnings Data Configuration}
\format{
A list containing configuration parameters for earnings data fetching.
}
\usage{
EARNINGS_CONFIG
}
\description{
Configuration object for fetching quarterly earnings data from Alpha Vantage.
This data provides timing metadata for when financial statements were reported.
}
\keyword{datasets}
