% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{ETF_PROFILE_CONFIG}
\alias{ETF_PROFILE_CONFIG}
\title{ETF Profile Data Configuration}
\format{
A list containing configuration parameters for ETF profile data fetching.
}
\usage{
ETF_PROFILE_CONFIG
}
\description{
Configuration object for fetching ETF profile data from Alpha Vantage.
}
\keyword{datasets}
