% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_artifacts_from_s3.R
\name{find_latest_artifact_date}
\alias{find_latest_artifact_date}
\title{Find Latest Artifact Date}
\usage{
find_latest_artifact_date(bucket_name, region)
}
\arguments{
\item{bucket_name}{character: S3 bucket name}

\item{region}{character: AWS region}
}
\value{
Date: Most recent artifact date
}
\description{
Finds the most recent artifact date in S3.
}
\keyword{internal}
