% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_options_term_structure.R
\name{find_latest_options_artifact_date}
\alias{find_latest_options_artifact_date}
\title{Find Latest Options Artifact Date in S3}
\usage{
find_latest_options_artifact_date(bucket_name, region)
}
\arguments{
\item{bucket_name}{character: S3 bucket name}

\item{region}{character: AWS region}
}
\value{
Date: Most recent options artifact date
}
\description{
Find Latest Options Artifact Date in S3
}
\keyword{internal}
