% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_batch_requests.R
\name{get_api_function_for_data_type}
\alias{get_api_function_for_data_type}
\title{Get Alpha Vantage API Function Name for Data Type}
\usage{
get_api_function_for_data_type(data_type)
}
\arguments{
\item{data_type}{character: One of "price", "splits", "balance_sheet",
"income_statement", "cash_flow", "earnings"}
}
\value{
character: The Alpha Vantage API function name
}
\description{
Get Alpha Vantage API Function Name for Data Type
}
\keyword{internal}
