% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_for_data_type.R
\name{get_config_for_data_type}
\alias{get_config_for_data_type}
\title{Get Config for Data Type}
\usage{
get_config_for_data_type(data_type)
}
\arguments{
\item{data_type}{character: Type of data (e.g., "balance_sheet", "price")}
}
\value{
list: Configuration object, or NULL if unknown data_type
}
\description{
Maps a data type name to its corresponding configuration object.
}
\keyword{internal}
