% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earnings_estimates_metrics.R
\name{get_earnings_estimates_metrics}
\alias{get_earnings_estimates_metrics}
\title{Get Earnings Estimates Metric Names}
\usage{
get_earnings_estimates_metrics()
}
\value{
character vector of earnings estimates metric names
}
\description{
Returns the column names for earnings estimates data.
}
\keyword{internal}
