% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spy_constituents.R
\name{get_spy_constituents}
\alias{get_spy_constituents}
\title{Get vector of tickers for S&P500 Constituents (as of July 2025)}
\usage{
get_spy_constituents()
}
\value{
A character vector of tickers in the S&P 500
}
\description{
Get vector of tickers for S&P500 Constituents (as of July 2025)
}
