% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{INCOME_STATEMENT_CONFIG}
\alias{INCOME_STATEMENT_CONFIG}
\title{Income Statement Data Configuration}
\format{
A list containing configuration parameters for income statement data fetching.
}
\usage{
INCOME_STATEMENT_CONFIG
}
\description{
Configuration object for fetching quarterly income statement data from Alpha Vantage.
}
\keyword{datasets}
