% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system2_with_timeout.R
\name{is_timeout_result}
\alias{is_timeout_result}
\title{Check if Result is a Timeout}
\usage{
is_timeout_result(result)
}
\arguments{
\item{result}{Result from system2_with_timeout()}
}
\value{
Logical TRUE if result was a timeout
}
\description{
Check if Result is a Timeout
}
\keyword{internal}
