% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_artifacts_from_s3.R
\name{load_price_artifact}
\alias{load_price_artifact}
\title{Load Price Artifact from S3}
\usage{
load_price_artifact(bucket_name, artifact_date = NULL, region = "us-east-1")
}
\arguments{
\item{bucket_name}{character: S3 bucket name}

\item{artifact_date}{Date: Date of the artifact (default: latest)}

\item{region}{character: AWS region (default: "us-east-1")}
}
\value{
tibble: Daily price data
}
\description{
Downloads and loads the price artifact from S3.
}
