% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_log_message.R
\name{log_error}
\alias{log_error}
\title{Log Error}
\usage{
log_error(ticker, error_msg)
}
\arguments{
\item{ticker}{Character. Ticker symbol}

\item{error_msg}{Character. Error message}
}
\value{
NULL (called for side effects)
}
\description{
Logs an error message with ticker context.
}
\keyword{internal}
