% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_api_request.R
\name{make_api_request}
\alias{make_api_request}
\title{Make API request to Alpha Vantage for daily adjusted price data}
\usage{
make_api_request(symbol, outputsize, datatype, api_key)
}
\arguments{
\item{symbol}{Character. The equity symbol}

\item{outputsize}{Character. Either "compact" or "full"}

\item{datatype}{Character. Either "json" or "csv"}

\item{api_key}{Character. Alpha Vantage API key}
}
\value{
Raw httr response object
}
\description{
Make API request to Alpha Vantage for daily adjusted price data
}
\keyword{internal}
