% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_etf_profile_request.R
\name{make_etf_profile_request}
\alias{make_etf_profile_request}
\title{Make API request to Alpha Vantage for ETF profile data}
\usage{
make_etf_profile_request(symbol, api_key = NULL)
}
\arguments{
\item{symbol}{Character. The ETF symbol}

\item{api_key}{Character. Alpha Vantage API key}
}
\value{
Raw httr response object
}
\description{
Make API request to Alpha Vantage for ETF profile data
}
\keyword{internal}
