% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_income_statement_request.R
\name{make_income_statement_request}
\alias{make_income_statement_request}
\title{Make API request to Alpha Vantage for income statement data}
\usage{
make_income_statement_request(ticker, api_key)
}
\arguments{
\item{ticker}{Character. The equity ticker}

\item{api_key}{Character. Alpha Vantage API key}
}
\value{
Raw httr response object
}
\description{
Make API request to Alpha Vantage for income statement data
}
\keyword{internal}
