% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{OVERVIEW_CONFIG}
\alias{OVERVIEW_CONFIG}
\title{Company Overview Data Configuration}
\format{
A list containing configuration parameters for company overview data fetching.
}
\usage{
OVERVIEW_CONFIG
}
\description{
Configuration object for fetching company metadata from Alpha Vantage.
}
\keyword{datasets}
