% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_api_response.R
\name{parse_api_response}
\alias{parse_api_response}
\title{Parse API response to tibble}
\usage{
parse_api_response(response, ticker, datatype)
}
\arguments{
\item{response}{Raw httr response object}

\item{ticker}{Character. The equity ticker}

\item{datatype}{Character. Either "json" or "csv"}
}
\value{
A tibble with daily adjusted price data
}
\description{
Parse API response to tibble
}
\keyword{internal}
