% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_historical_options_response.R
\name{parse_historical_options_response}
\alias{parse_historical_options_response}
\title{Parse Historical Options Response to Tibble}
\usage{
parse_historical_options_response(response, ticker, datatype = "csv")
}
\arguments{
\item{response}{httr2 response object}

\item{ticker}{Character. The equity ticker}

\item{datatype}{Character. Either "json" or "csv"}
}
\value{
A tibble with option chain data
}
\description{
Parse Historical Options Response to Tibble
}
