% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_price_response.R
\name{parse_price_response}
\alias{parse_price_response}
\title{Parse Price Response to Tibble}
\usage{
parse_price_response(response, ticker, datatype)
}
\arguments{
\item{response}{httr2 response object}

\item{ticker}{Character. The equity ticker}

\item{datatype}{Character. Either "json" or "csv"}
}
\value{
A tibble with daily adjusted price data
}
\description{
Parse Price Response to Tibble
}
\keyword{internal}
