% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vantage_configs.R
\docType{data}
\name{SPLITS_CONFIG}
\alias{SPLITS_CONFIG}
\title{Corporate Actions - Splits Data Configuration}
\format{
A list containing configuration parameters for splits data fetching.
}
\usage{
SPLITS_CONFIG
}
\description{
Configuration object for fetching historical stock split events from Alpha Vantage.
}
\keyword{datasets}
