% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_df_type.R
\name{validate_df_type}
\alias{validate_df_type}
\title{Validate DataFrame Type}
\usage{
validate_df_type(data)
}
\arguments{
\item{data}{Object to validate as a data.frame}
}
\value{
NULL (called for side effects)
}
\description{
Validates that an object is a data.frame.
This function is called for its side effects and will stop execution with
an error message if validation fails.
}
\keyword{internal}
